Cleared xen_diag interface
