+++ templates/xen/debug/mods.config
@@ -14,8 +14,8 @@ configuration conf {
 	include embox.kernel.critical
 	include embox.mem.NoPage
 
-	@Runlevel(2) include embox.driver.diag(impl="embox__driver__diag__mem_diag")
-	include embox.driver.diag.mem_diag
+	@Runlevel(2) include embox.driver.diag(impl="embox__driver__diag__xen_diag")
+	include embox.driver.diag.xen_diag
 
 
 	@Runlevel(2) include embox.util.LibUtil
